+++ enterprise/internal/codeintel/autoindexing/service.go
@@ -3,21 +3,17 @@ package autoindexing
 import (
 	"context"
 	"fmt"
-	"os"
 	"time"
 
 	"github.com/grafana/regexp"
 	otlog "github.com/opentracing/opentracing-go/log"
 	"github.com/sourcegraph/log"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/autoindexing/internal/inference"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/autoindexing/internal/store"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/autoindexing/shared"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/shared/types"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/env"
-	"github.com/sourcegraph/sourcegraph/internal/errcode"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/symbols"
@@ -27,16 +23,14 @@ import (
 )
 
 type Service struct {
-	store store.Store
-
+	store           store.Store
 	uploadSvc       UploadService
 	inferenceSvc    InferenceService
 	repoUpdater     RepoUpdaterClient
 	gitserverClient GitserverClient
 	symbolsClient   *symbols.Client
-
-	logger     log.Logger
-	operations *operations
+	logger          log.Logger
+	operations      *operations
 }
 
 func newService(
@@ -49,29 +43,15 @@ func newService(
 	observationContext *observation.Context,
 ) *Service {
 	return &Service{
-		store: store,
-
+		store:           store,
 		uploadSvc:       uploadSvc,
 		inferenceSvc:    inferenceSvc,
 		repoUpdater:     repoUpdater,
 		gitserverClient: gitserver,
 		symbolsClient:   symbolsClient,
-
-		logger:     observationContext.Logger,
-		operations: newOperations(observationContext),
+		logger:          observationContext.Logger,
+		operations:      newOperations(observationContext),
 	}
-}
-
-type indexEnqueuer struct {
-	store           store.Store
-	uploadSvc       UploadService
-	inferenceSvc    InferenceService
-	repoUpdater     RepoUpdaterClient
-	gitserverClient GitserverClient
-	symbolsClient   *symbols.Client
-	logger          log.Logger
-	operations      *operations
-	inferer         *inferer
 }
 
 func (s *Service) IndexEnqueuer() IndexEnqueuer {
@@ -85,6 +65,19 @@ func (s *Service) IndexEnqueuer() IndexEnqueuer {
 		logger:          s.logger,
 		operations:      s.operations,
 		inferer:         s.inferer(),
+	}
+}
+
+func (s *Service) inferer() *inferer {
+	return &inferer{
+		store:           s.store,
+		uploadSvc:       s.uploadSvc,
+		inferenceSvc:    s.inferenceSvc,
+		repoUpdater:     s.repoUpdater,
+		gitserverClient: s.gitserverClient,
+		symbolsClient:   s.symbolsClient,
+		logger:          s.logger,
+		operations:      s.operations,
 	}
 }
 
@@ -294,377 +287,14 @@ func (s *Service) QueueIndexes(ctx context.Context, repositoryID int, rev, confi
 	return s.IndexEnqueuer().QueueIndexes(ctx, repositoryID, rev, configuration, force, bypassLimit)
 }
 
-// QueueIndexes enqueues a set of index jobs for the following repository and commit. If a non-empty
-// configuration is given, it will be used to determine the set of jobs to enqueue. Otherwise, it will
-// the configuration will be determined based on the regular index scheduling rules: first read any
-// in-repo configuration (e.g., sourcegraph.yaml), then look for any existing in-database configuration,
-// finally falling back to the automatically inferred configuration based on the repo contents at the
-// target commit.
-//
-// If the force flag is false, then the presence of an upload or index record for this given repository and commit
-// will cause this method to no-op. Note that this is NOT a guarantee that there will never be any duplicate records
-// when the flag is false.
-func (s *indexEnqueuer) QueueIndexes(ctx context.Context, repositoryID int, rev, configuration string, force, bypassLimit bool) (_ []types.Index, err error) {
-	ctx, trace, endObservation := s.operations.queueIndex.With(ctx, &err, observation.Args{
-		LogFields: []otlog.Field{
-			otlog.Int("repositoryID", repositoryID),
-			otlog.String("rev", rev),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	commitID, err := s.gitserverClient.ResolveRevision(ctx, repositoryID, rev)
-	if err != nil {
-		return nil, errors.Wrap(err, "gitserver.ResolveRevision")
-	}
-	commit := string(commitID)
-	trace.Log(otlog.String("commit", commit))
-
-	return s.queueIndexForRepositoryAndCommit(ctx, repositoryID, commit, configuration, force, bypassLimit, nil) // trace)
-}
-
 func (s *Service) QueueIndexesForPackage(ctx context.Context, pkg precise.Package) (err error) {
 	return s.IndexEnqueuer().QueueIndexesForPackage(ctx, pkg)
 }
-
-// QueueIndexesForPackage enqueues index jobs for a dependency of a recently-processed precise code
-// intelligence index.
-func (s *indexEnqueuer) QueueIndexesForPackage(ctx context.Context, pkg precise.Package) (err error) {
-	ctx, trace, endObservation := s.operations.queueIndexForPackage.With(ctx, &err, observation.Args{
-		LogFields: []otlog.Field{
-			otlog.String("scheme", pkg.Scheme),
-			otlog.String("name", pkg.Name),
-			otlog.String("version", pkg.Version),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	repoName, revision, ok := inference.InferRepositoryAndRevision(pkg)
-	if !ok {
-		return nil
-	}
-	trace.Log(otlog.String("repoName", string(repoName)))
-	trace.Log(otlog.String("revision", revision))
-
-	resp, err := s.repoUpdater.EnqueueRepoUpdate(ctx, repoName)
-	if err != nil {
-		if errcode.IsNotFound(err) {
-			return nil
-		}
-
-		return errors.Wrap(err, "repoUpdater.EnqueueRepoUpdate")
-	}
-
-	commit, err := s.gitserverClient.ResolveRevision(ctx, int(resp.ID), revision)
-	if err != nil {
-		if errcode.IsNotFound(err) {
-			return nil
-		}
-
-		return errors.Wrap(err, "gitserverClient.ResolveRevision")
-	}
-
-	_, err = s.queueIndexForRepositoryAndCommit(ctx, int(resp.ID), string(commit), "", false, false, nil) // trace)
-	return err
-}
-
-var (
-	overrideScript                           = os.Getenv("SRC_CODEINTEL_INFERENCE_OVERRIDE_SCRIPT")
-	maximumIndexJobsPerInferredConfiguration = env.MustGetInt("PRECISE_CODE_INTEL_AUTO_INDEX_MAXIMUM_INDEX_JOBS_PER_INFERRED_CONFIGURATION", 25, "Repositories with a number of inferred auto-index jobs exceeding this threshold will not be auto-indexed.")
-)
 
 func (s *Service) InferIndexJobsFromRepositoryStructure(ctx context.Context, repositoryID int, commit string, bypassLimit bool) ([]config.IndexJob, error) {
 	return s.inferer().InferIndexJobsFromRepositoryStructure(ctx, repositoryID, commit, bypassLimit)
 }
 
-// InferIndexJobsFromRepositoryStructure collects the result of  InferIndexJobs over all registered recognizers.
-func (s *inferer) InferIndexJobsFromRepositoryStructure(ctx context.Context, repositoryID int, commit string, bypassLimit bool) ([]config.IndexJob, error) {
-	repoName, err := s.uploadSvc.GetRepoName(ctx, repositoryID)
-	if err != nil {
-		return nil, err
-	}
-
-	script, err := s.store.GetInferenceScript(ctx)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to fetch inference script from database")
-	}
-	if script == "" {
-		script = overrideScript
-	}
-
-	indexes, err := s.inferenceSvc.InferIndexJobs(ctx, api.RepoName(repoName), commit, script)
-	if err != nil {
-		return nil, err
-	}
-
-	if !bypassLimit && len(indexes) > maximumIndexJobsPerInferredConfiguration {
-		s.logger.Info("Too many inferred roots. Scheduling no index jobs for repository.", log.Int("repository_id", repositoryID))
-		return nil, nil
-	}
-
-	return indexes, nil
-}
-
 func (s *Service) InferIndexJobHintsFromRepositoryStructure(ctx context.Context, repositoryID int, commit string) ([]config.IndexJobHint, error) {
 	return s.inferer().InferIndexJobHintsFromRepositoryStructure(ctx, repositoryID, commit)
-}
-
-// inferIndexJobsFromRepositoryStructure collects the result of  InferIndexJobHints over all registered recognizers.
-func (s *inferer) InferIndexJobHintsFromRepositoryStructure(ctx context.Context, repositoryID int, commit string) ([]config.IndexJobHint, error) {
-	repoName, err := s.uploadSvc.GetRepoName(ctx, repositoryID)
-	if err != nil {
-		return nil, err
-	}
-
-	indexes, err := s.inferenceSvc.InferIndexJobHints(ctx, api.RepoName(repoName), commit, overrideScript)
-	if err != nil {
-		return nil, err
-	}
-
-	return indexes, nil
-}
-
-// queueIndexForRepositoryAndCommit determines a set of index jobs to enqueue for the given repository and commit.
-//
-// If the force flag is false, then the presence of an upload or index record for this given repository and commit
-// will cause this method to no-op. Note that this is NOT a guarantee that there will never be any duplicate records
-// when the flag is false.
-func (s *indexEnqueuer) queueIndexForRepositoryAndCommit(ctx context.Context, repositoryID int, commit, configuration string, force, bypassLimit bool, trace observation.TraceLogger) ([]types.Index, error) {
-	if !force {
-		isQueued, err := s.store.IsQueued(ctx, repositoryID, commit)
-		if err != nil {
-			return nil, errors.Wrap(err, "dbstore.IsQueued")
-		}
-		if isQueued {
-			return nil, nil
-		}
-	}
-
-	indexes, err := s.inferer.getIndexRecords(ctx, repositoryID, commit, configuration, bypassLimit)
-	if err != nil {
-		return nil, err
-	}
-	if len(indexes) == 0 {
-		return nil, nil
-	}
-
-	indexesToInsert := indexes
-	if !force {
-		indexesToInsert = []types.Index{}
-		for _, index := range indexes {
-			isQueued, err := s.store.IsQueuedRootIndexer(ctx, repositoryID, commit, index.Root, index.Indexer)
-			if err != nil {
-				return nil, errors.Wrap(err, "dbstore.IsQueuedRootIndexer")
-			}
-			if !isQueued {
-				indexesToInsert = append(indexesToInsert, index)
-			}
-		}
-	}
-
-	return s.store.InsertIndexes(ctx, indexesToInsert)
-}
-
-type inferer struct {
-	store           store.Store
-	uploadSvc       UploadService
-	inferenceSvc    InferenceService
-	repoUpdater     RepoUpdaterClient
-	gitserverClient GitserverClient
-	symbolsClient   *symbols.Client
-	logger          log.Logger
-	operations      *operations
-}
-
-func (s *Service) inferer() *inferer {
-	return &inferer{
-		store:           s.store,
-		uploadSvc:       s.uploadSvc,
-		inferenceSvc:    s.inferenceSvc,
-		repoUpdater:     s.repoUpdater,
-		gitserverClient: s.gitserverClient,
-		symbolsClient:   s.symbolsClient,
-		logger:          s.logger,
-		operations:      s.operations,
-	}
-}
-
-type configurationFactoryFunc func(ctx context.Context, repositoryID int, commit string, bypassLimit bool) ([]types.Index, bool, error)
-
-// getIndexRecords determines the set of index records that should be enqueued for the given commit.
-// For each repository, we look for index configuration in the following order:
-//
-//   - supplied explicitly via parameter
-//   - in the database
-//   - committed to `sourcegraph.yaml` in the repository
-//   - inferred from the repository structure
-func (s *inferer) getIndexRecords(ctx context.Context, repositoryID int, commit, configuration string, bypassLimit bool) ([]types.Index, error) {
-	fns := []configurationFactoryFunc{
-		makeExplicitConfigurationFactory(configuration),
-		s.getIndexRecordsFromConfigurationInDatabase,
-		s.getIndexRecordsFromConfigurationInRepository,
-		s.inferIndexRecordsFromRepositoryStructure,
-	}
-
-	for _, fn := range fns {
-		if indexRecords, ok, err := fn(ctx, repositoryID, commit, bypassLimit); err != nil {
-			return nil, err
-		} else if ok {
-			return indexRecords, nil
-		}
-	}
-
-	return nil, nil
-}
-
-// makeExplicitConfigurationFactory returns a factory that returns a set of index jobs configured
-// explicitly via a GraphQL query parameter. If no configuration was supplield then a false valued
-// flag is returned.
-func makeExplicitConfigurationFactory(configuration string) configurationFactoryFunc {
-	logger := log.Scoped("explicitConfigurationFactory", "")
-	return func(ctx context.Context, repositoryID int, commit string, _ bool) ([]types.Index, bool, error) {
-		if configuration == "" {
-			return nil, false, nil
-		}
-
-		indexConfiguration, err := config.UnmarshalJSON([]byte(configuration))
-		if err != nil {
-			// We failed here, but do not try to fall back on another method as having
-			// an explicit config supplied via parameter should always take precedence,
-			// even if it's broken.
-			logger.Warn("Failed to unmarshal index configuration", log.Int("repository_id", repositoryID), log.Error(err))
-			return nil, true, nil
-		}
-
-		return convertIndexConfiguration(repositoryID, commit, indexConfiguration), true, nil
-	}
-}
-
-// getIndexRecordsFromConfigurationInDatabase returns a set of index jobs configured via the UI for
-// the given repository. If no jobs are configured via the UI then a false valued flag is returned.
-func (s *inferer) getIndexRecordsFromConfigurationInDatabase(ctx context.Context, repositoryID int, commit string, _ bool) ([]types.Index, bool, error) {
-	indexConfigurationRecord, ok, err := s.store.GetIndexConfigurationByRepositoryID(ctx, repositoryID)
-	if err != nil {
-		return nil, false, errors.Wrap(err, "dbstore.GetIndexConfigurationByRepositoryID")
-	}
-	if !ok {
-		return nil, false, nil
-	}
-
-	indexConfiguration, err := config.UnmarshalJSON(indexConfigurationRecord.Data)
-	if err != nil {
-		// We failed here, but do not try to fall back on another method as having
-		// an explicit config in the database should always take precedence, even
-		// if it's broken.
-		s.logger.Warn("Failed to unmarshal index configuration", log.Int("repository_id", repositoryID), log.Error(err))
-		return nil, true, nil
-	}
-
-	return convertIndexConfiguration(repositoryID, commit, indexConfiguration), true, nil
-}
-
-// getIndexRecordsFromConfigurationInRepository returns a set of index jobs configured via a committed
-// configuration file at the given commit. If no jobs are configured within the repository then a false
-// valued flag is returned.
-func (s *inferer) getIndexRecordsFromConfigurationInRepository(ctx context.Context, repositoryID int, commit string, _ bool) ([]types.Index, bool, error) {
-	isConfigured, err := s.gitserverClient.FileExists(ctx, repositoryID, commit, "sourcegraph.yaml")
-	if err != nil {
-		return nil, false, errors.Wrap(err, "gitserver.FileExists")
-	}
-	if !isConfigured {
-		return nil, false, nil
-	}
-
-	content, err := s.gitserverClient.RawContents(ctx, repositoryID, commit, "sourcegraph.yaml")
-	if err != nil {
-		return nil, false, errors.Wrap(err, "gitserver.RawContents")
-	}
-
-	indexConfiguration, err := config.UnmarshalYAML(content)
-	if err != nil {
-		// We failed here, but do not try to fall back on another method as having
-		// an explicit config in the repository should always take precedence over
-		// an auto-inferred configuration, even if it's broken.
-		s.logger.Warn("Failed to unmarshal index configuration", log.Int("repository_id", repositoryID), log.Error(err))
-		return nil, true, nil
-	}
-
-	return convertIndexConfiguration(repositoryID, commit, indexConfiguration), true, nil
-}
-
-// inferIndexRecordsFromRepositoryStructure looks at the repository contents at the given commit and
-// determines a set of index jobs that are likely to succeed. If no jobs could be inferred then a
-// false valued flag is returned.
-func (s *inferer) inferIndexRecordsFromRepositoryStructure(ctx context.Context, repositoryID int, commit string, bypassLimit bool) ([]types.Index, bool, error) {
-	indexJobs, err := s.InferIndexJobsFromRepositoryStructure(ctx, repositoryID, commit, bypassLimit)
-	if err != nil || len(indexJobs) == 0 {
-		return nil, false, err
-	}
-
-	return convertInferredConfiguration(repositoryID, commit, indexJobs), true, nil
-}
-
-// convertIndexConfiguration converts an index configuration object into a set of index records to be
-// inserted into the database.
-func convertIndexConfiguration(repositoryID int, commit string, indexConfiguration config.IndexConfiguration) (indexes []types.Index) {
-	for _, indexJob := range indexConfiguration.IndexJobs {
-		var dockerSteps []types.DockerStep
-		for _, dockerStep := range indexConfiguration.SharedSteps {
-			dockerSteps = append(dockerSteps, types.DockerStep{
-				Root:     dockerStep.Root,
-				Image:    dockerStep.Image,
-				Commands: dockerStep.Commands,
-			})
-		}
-		for _, dockerStep := range indexJob.Steps {
-			dockerSteps = append(dockerSteps, types.DockerStep{
-				Root:     dockerStep.Root,
-				Image:    dockerStep.Image,
-				Commands: dockerStep.Commands,
-			})
-		}
-
-		indexes = append(indexes, types.Index{
-			Commit:       commit,
-			RepositoryID: repositoryID,
-			State:        "queued",
-			DockerSteps:  dockerSteps,
-			LocalSteps:   indexJob.LocalSteps,
-			Root:         indexJob.Root,
-			Indexer:      indexJob.Indexer,
-			IndexerArgs:  indexJob.IndexerArgs,
-			Outfile:      indexJob.Outfile,
-		})
-	}
-
-	return indexes
-}
-
-// convertInferredConfiguration converts a set of index jobs into a set of index records to be inserted
-// into the database.
-func convertInferredConfiguration(repositoryID int, commit string, indexJobs []config.IndexJob) (indexes []types.Index) {
-	for _, indexJob := range indexJobs {
-		var dockerSteps []types.DockerStep
-		for _, dockerStep := range indexJob.Steps {
-			dockerSteps = append(dockerSteps, types.DockerStep{
-				Root:     dockerStep.Root,
-				Image:    dockerStep.Image,
-				Commands: dockerStep.Commands,
-			})
-		}
-
-		indexes = append(indexes, types.Index{
-			RepositoryID: repositoryID,
-			Commit:       commit,
-			State:        "queued",
-			DockerSteps:  dockerSteps,
-			LocalSteps:   indexJob.LocalSteps,
-			Root:         indexJob.Root,
-			Indexer:      indexJob.Indexer,
-			IndexerArgs:  indexJob.IndexerArgs,
-			Outfile:      indexJob.Outfile,
-		})
-	}
-
-	return indexes
 }
